Finish core ESM migration
