+++ docs/plugins.md
@@ -13,11 +13,11 @@ npm install @postr/syndicator-instagram
 ```
 
 ```js
-const Postr = require('@postr/core')
-const InstagramSyndicator = require('@postr/syndicator-instagram')
+import { postr } from '@postr/core'
+import { InstagramSyndicator } from '@postr/syndicator-instagram'
 
-const postr = new Postr(options)
-postr.use(InstagramSyndicator, {
+const endpoint = postr(options)
+endpoint.use(InstagramSyndicator, {
   username: 'instagramusername',
   password: 'instagrampassword',
   default: true, // Syndicate to instagram even if no syndication is set
@@ -32,11 +32,11 @@ npm install @postr/syndicator-superfeedr
 ```
 
 ```js
-const Postr = require('@postr/core')
-const SuperfeedrSyndicator = require('@postr/syndicator-superfeedr')
+import { postr } from '@postr/core'
+import { SuperfeedrSyndicator } from '@postr/syndicator-superfeedr'
 
-const postr = new Postr(options)
-postr.use(SuperfeedrSyndicator, {
+const endpoint = postr(options)
+endpoint.use(SuperfeedrSyndicator, {
   hub: 'superfeedrhub',
   domain: 'superfeedrdomain',
   default: true, // Syndicate to Superfeedr even if no syndication is set
@@ -50,11 +50,11 @@ npm install @postr/plugin-webmention-endpoint
 ```
 
 ```js
-const Postr = require('@postr/core')
-const WebmentionEndpoint = require('@postr/plugin-webmention-endpoint')
+import { postr } from '@postr/core'
+import { WebmentionEndpoint } from '@postr/plugin-webmention-endpoint'
 
-const postr = new Postr(options)
-postr.use(WebmentionEndpoint, {
+const endpoint = postr(options)
+endpoint.use(WebmentionEndpoint, {
   notifier: function(notification) {
     // Handle sending notifications somehow.
     // The notification variable can be a simple string or a mf2 object
@@ -71,7 +71,8 @@ npm install @postr/plugin
 ```
 
 ```js
-const PostrPlugin = require('@postr/plugin')
+import { PostrPlugin } from '@postr/plugin'
+import { Router } from 'express'
 
 class MyPlugin extends PostrPlugin {
   // Plugins must have a constuctor that sets at least an id and name and any other default options
@@ -98,7 +99,7 @@ class MyPlugin extends PostrPlugin {
 
 
   router() {
-    const router = require('express').Router()
+    const router = Router()
 
     router.get('/', (req, res, next) => {
       // Handle router request however you wish
@@ -122,7 +123,7 @@ class MyPlugin extends PostrPlugin {
   }
 }
 
-module.exports = MyPlugin
+export default MyPlugin
 ```
 
 ### Developing a syndication plugin
@@ -134,10 +135,9 @@ npm install @postr/syndicator
 ```
 
 ```js
-const axios = require('axios')
-const BaseSyndicator = require('@postr/syndicator')
+import { Syndicator } from '@postr/syndicator'
 
-class MySyndicator extends BaseSyndicator {
+class MySyndicator extends Syndicator {
   // Constructor that sets at least an id, name and if the syndication should be considered default or not
   constructor({ options, imports }) {
     options = Object.assign(
@@ -172,6 +172,6 @@ class MySyndicator extends BaseSyndicator {
   }
 }
 
-module.exports = MySyndicator
+export default MySyndicator
 
 ```
