--- packages/core/lib/db-middleware/create-refs.js
@@ -1,42 +1,0 @@
-const { isRxDocument } = require('rxdb')
-const { getUrlsFromMf2, getHEntry, config } = globalThis.__postr
-const siteBaseUrl = config.get('siteBaseUrl')
-
-// This will only work with an rxdoc, should be run in parallel to prevent it slowing down other databse stuff
-
-/**
- * Scrapes urls from a RxDocument and saves the mf2 data in a _refs array
- * @param {RxDocument} doc The document to add references to
- */
-module.exports = async post => {
-  if (isRxDocument(post)) {
-    // Don't update rx docs
-    return doc
-  }
-  const urls = getUrlsFromMf2(post)
-  if (urls && urls.length) {
-    // Got the urls to scrape
-    for (const url of urls) {
-      if (!url.startsWith(siteBaseUrl)) {
-        // This is not a url to a page on your site, so lets scrape it
-        try {
-          const res = await getHEntry(url)
-          if (res) {
-            // NOTE: I store this as an array rather than the references object specified by jf2 as dots in property names cause too many issues
-            if (!post.cms) {
-              post.cms = { references: [] }
-            }
-            if (!post.cms.references) {
-              post.cms.references = []
-            }
-            post.cms.references.push(res)
-          }
-        } catch (err) {
-          // Error getting h-entry from url. Probably not a big deal
-          console.warn(`[Error getting h-entry from ${url}]`, err)
-        }
-      }
-    }
-  }
-  return post
-}
