--- packages/core/lib/middlewear/media-endpoint.js
@@ -1,42 +1,0 @@
-const hasScope = require('./has-scope')
-const { saveFile, placeholders } = globalThis.__postr
-
-/**
- * Micropub media endpoint as an express middlewear
- * @param {object} req Express request
- * @param {object} res Express response
- * @param {function} next Express next function
- */
-module.exports = [
-  hasScope('create'),
-  (req, res, next) => {
-    if (!req.file || req.file.truncated || !req.file.buffer) {
-      res.status(400)
-      return res.json({
-        error: 'Error with file',
-        error_description: 'Error or missing file',
-      })
-    }
-
-    const filename = req.file.originalname
-    // Save file
-    saveFile(req.file.buffer, filename)
-      .then(url => {
-        url = placeholders.replace(url)
-        res.status(201)
-        res.header('Location', url)
-        return res.json({
-          location: url,
-        })
-      })
-      .catch(err => {
-        console.log('Error writing media endpoint file', err)
-        res.status(500)
-        return res.json({
-          error: 'Error saving file',
-          error_description:
-            'There was an error writing the media file to disk',
-        })
-      })
-  },
-]
