+++ packages/core/src/index.mts
@@ -1,37 +1,20 @@
 /**
  * @postr/core — ESM + TypeScript entrypoint
  *
- * Wave 8: Genuine ESM leaf modules with globalThis bridge for CJS spine.
+ * Wave 9: All modules are native ESM. No globalThis bridge needed.
  *
  * ## CJS interop strategy
  *
- * **Wave 8 modules (this file's direct imports)**
- *   - Compiled to ESM (`.mjs`), imported directly.
- *   - Includes: config, generateSearch, post-type-discovery, placeholders,
- *     save-file, save-file-from-url, get-hentry, error, is-node,
- *     get-permalink-from-mf2, get-urls-from-mf2, today-media-path,
- *     append-to-filename, statics, schema modules.
- *
- * **Wave 9 CJS spine (loaded via createRequire)**
- *   - Copied from `lib/` as `.js` files during build.
- *   - Loaded via `createRequire` from this ESM context.
- *   - Includes: index.js (deleted — logic inlined here), db.js, plugins.js,
- *     router.js, middlewear/, db-middleware/.
- *   - These CJS files access Wave 8 ESM modules via globalThis.__postr,
- *     populated before any CJS require runs.
- *
  * **RxDB / PouchDB** (CJS-only packages)
- *   - All loaded via createRequire from the ESM context.
+ *   - All loaded via createRequire inside the consuming module (db.mjs, plugins.mjs).
+ *
+ * **Other CJS deps** (file-type, image-size, sharp, ajv, send-webmention, etc.)
+ *   - Loaded via createRequire inside the consuming module.
  *
  * **Express and friends**
- *   - CJS packages but generally ESM-safe in Node 24.
- *   - Loaded via createRequire for consistency.
+ *   - CJS packages but generally ESM-safe in Node 24 via default interop.
+ *   - Imported directly with ESM `import` syntax.
  */
-
-import { createRequire } from 'node:module'
-
-// ===== ESM bridge — populated before CJS spine loads =====
-;(globalThis as any).__postr = {}
 
 // Wave 8 — genuine ESM leaf modules
 import * as config from './lib/config.mjs'
@@ -41,51 +24,10 @@ import {
   getAvailablePostTypes,
   addPostType,
 } from './lib/post-type-discovery.mjs'
-import * as placeholders from './lib/placeholders.mjs'
-import saveFileFn from './lib/save-file.mjs'
-import saveFileFromUrlFn from './lib/save-file-from-url.mjs'
-import getHEntryFn from './lib/get-hentry.mjs'
-import MicropubError from './lib/error.mjs'
-import isNode from './lib/is-node.mjs'
-import getPermalinkFromMf2 from './lib/get-permalink-from-mf2.mjs'
-import getUrlsFromMf2 from './lib/get-urls-from-mf2.mjs'
-import todayMediaPathFn from './lib/today-media-path.mjs'
-import appendToFilenameFn from './lib/append-to-filename.mjs'
-import schema from './schema/base.mjs'
-import migrationStrategies from './schema/migration-strategies.mjs'
-import getPostSchema from './schema/postTypes.mjs'
-import * as staticMethods from './lib/statics/index.mjs'
-
-// Populate globalThis bridge before CJS spine is loaded
-const bridge = (globalThis as any).__postr
-bridge.config = config
-bridge.generateSearch = generateSearchFn
-bridge.getPostType = getPostType
-bridge.getAvailablePostTypes = getAvailablePostTypes
-bridge.addPostType = addPostType
-bridge.placeholders = placeholders
-bridge.saveFile = saveFileFn
-bridge.saveFileFromUrl = saveFileFromUrlFn
-bridge.getHEntry = getHEntryFn
-bridge.MicropubError = MicropubError
-bridge.isNode = isNode
-bridge.getPermalinkFromMf2 = getPermalinkFromMf2
-bridge.getUrlsFromMf2 = getUrlsFromMf2
-bridge.todayMediaPath = todayMediaPathFn
-bridge.appendToFilename = appendToFilenameFn
-bridge.schema = schema
-bridge.migrationStrategies = migrationStrategies
-bridge.getPostSchema = getPostSchema
-bridge.staticMethods = staticMethods
-bridge.getPermalink = getPermalinkFromMf2
-bridge.getUrls = getUrlsFromMf2
-bridge.replaceMf2 = placeholders.replaceMf2
-
-// ===== Wave 9 CJS spine =====
-const require = createRequire(import.meta.url)
-const { use, plugins } = require('./lib/plugins.js')
-const micropubRouter = require('./lib/router.js')
-const Collection = require('./lib/db.js')
+// Wave 9 — converted ESM spine modules
+import { use, plugins } from './lib/plugins.mjs'
+import { default as micropubRouter } from './lib/router.mjs'
+import { get as getCollection } from './lib/db.mjs'
 
 // ===== Public API types =====
 
@@ -95,7 +37,11 @@ export interface PostrCore {
   getCollection: () => Promise<unknown>
   router: unknown
   generateSearch: (text: string) => string
-  addPostType: (options: { id: string; name: string; discovery: (post: any) => boolean }) => void
+  addPostType: (options: {
+    id: string
+    name: string
+    discovery: (post: any) => boolean
+  }) => void
   micropubEndpoint: string
   mediaEndpoint: string
 }
@@ -129,7 +75,7 @@ export function postr(options: Record<string, unknown> = {}): PostrCore {
     return {
       use,
       plugins,
-      getCollection: Collection.get,
+      getCollection,
       router: micropubRouter,
       generateSearch: generateSearchFn,
       addPostType: addPostType,
